added function Cell in separated file
